requirements.txt
fastapi[standard]
ccxt
openai>=1.90.0
langchain>=0.3.26
langchain-openai>=0.3.24
python-dotenv>=1.1.0